Various Updates
Changed the proton mail aur package to a better supported one and added the flatpak as an install option. Changed the proton vpn to the official pacman package in the extra repository. Removed skype install option since its being discontinued in May 2025. Added Mission Center and Fastfetch as install options.